+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.16.0] - 2022-04-23
+
+### Added
+
+- Add optional holidays for Manitoba to the API
+
+### Updated
+
+- Add optional holidays to UI for Manitoba
+
 ## [3.15.0] - 2022-04-23
 
 ### Added